Clean up the repo for review
